Added a check for immunity for discord invite links

+++ index.js
@@ -157,7 +157,7 @@ client.on("message", msg => {
         }
         // Rules
         if (!msg.author.bot && msg.guild.id == config.channels.guild) {
-            if (!config.channels.ignore.includes(msg.channel.id)) {
+            if (!config.channels.ignore.includes(msg.channel.id) && !msg.author.roles.cache.some(role => role.id === config.roles.immuneRole)) {
                 if (msg.content.match(/(https?:\/\/)?(www\.)?(discord\.(gg|io|me|li)|discordapp\.com\/invite)\/.+[a-z]/gm)) {
                     if (config.automation.discordInvites.deleteMessage) {
                         msg.delete();
